Add a demo of what an enhanced help system for curl might look like
This script demonstrates what help output might look like if it were
broken down into multiple categories so only a few related options are
displayed at the same time rather than all 231 at once.

The curlh script works just like curl but has different help options.
run "curlh --help-demo" to see which options are available and
"curlh --help" to see what ones might be available in the full version.
Only a few of the possible categories are implemented here but it should
be enough to get a flavour of what a full implementation might be like.